src/main/java/com/thoughtworks/Item.java
package com.thoughtworks;

/**
 * Created by lyzhang on 2/25/16.
 */
public class Item {
    private String barCode;
    private String name;
    private Double price;
    private Integer count;

    public void setBarCode(String barCode) {
        this.barCode = barCode;
    }

    public String getBarCode() {
        return barCode;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public Double getPrice() {
        return price;
    }

    public void setPrice(Double price) {
        this.price = price;
    }

    public Integer getCount() {
        return count;
    }

    public void setCount(Integer count) {
        this.count = count;
    }
}
